requirements.txt
streamlit
openai
PyPDF2
pydub
python-docx
exceptions